curve: Refactor and move curve containers to subfolder.
