+++ libopenage/curve/CMakeLists.txt
@@ -1,19 +1,13 @@
 add_sources(libopenage
-	array.cpp
 	base_curve.cpp
 	continuous.cpp
 	discrete.cpp
 	discrete_mod.cpp
-    element_wrapper.cpp
 	interpolated.cpp
-	iterator.cpp
 	keyframe.cpp
 	keyframe_container.cpp
-	map.cpp
-	map_filter_iterator.cpp
-	queue.cpp
-	queue_filter_iterator.cpp
 	segmented.cpp
 )
 
+add_subdirectory("container")
 add_subdirectory("tests")
