+++ libopenage/curve/container/CMakeLists.txt
@@ -1,0 +1,9 @@
+add_sources(libopenage
+	array.cpp
+    element_wrapper.cpp
+	iterator.cpp
+	map.cpp
+	map_filter_iterator.cpp
+	queue.cpp
+	queue_filter_iterator.cpp
+)
