+++ libopenage/curve/container/map_filter_iterator.cpp
@@ -1,4 +1,4 @@
-// Copyright 2017-2018 the openage authors. See copying.md for legal info.
+// Copyright 2017-2025 the openage authors. See copying.md for legal info.
 
 #include "map_filter_iterator.h"
